chore: Extract Block from reader/mod.rs into reader/block.rs
